PIMD

PIMD.ExternalPotentialPropagator

PIMD.GeneralizedCayleyPropagator

PIMD.Integrator

PIMD.KineticPropagator

PIMD.LangevinPropagator

PIMD.Normal

PIMD.OneDimensionalQuantumSystem

PIMD.PotentialPropagator

PIMD.Primitive

PIMD.Staging

PIMD.estimatorHessian

PIMD.estimatorKinetic

PIMD.estimatorKineticVirial

PIMD.estimatorPotential

PIMD.md

PIMD.ExternalPotentialPropagator — Type

ExternalKineticPropagator(Δt)

Creates a propagator that updates momenta according to the canonical equation, with PIMD external potential only

$\boldsymbol p\leftarrow\boldsymbol p-\frac{\partial V_{\rm ext}}{\partial\boldsymbol q}\Delta t$

PIMD.GeneralizedCayleyPropagator — Type

GeneralizedCayleyPropagator(system, generalizedCayleyFunction, Δt)

Creates a generalized Cayley propagator that PIMD internal potential using a generalized cayley function $\varphi$

$\boldsymbol z\leftarrow\varphi(\mathsf J\mathsf H)\boldsymbol z$

Examples of generalized Cayley functions are exp(), cayley() and sqrtcayley() (recommended).

PIMD.Integrator — Type

Integrator(propagators...)

Constructs an integrator from a series of propagator, so that within a time step these propagators act sequentially on the phase.

PIMD.KineticPropagator — Type

KineticPropagator(Δt)

Creates a propagator that updates positions according to the canonical equation

$\boldsymbol q\leftarrow\boldsymbol q+\frac{\partial H}{\partial\boldsymbol p}\Delta t$

PIMD.LangevinPropagator — Type

LangevinPropagator(system, Δt)

Creates a propagator that thermostats momenta using Langevin equation

$\boldsymbol p\leftarrow e^{-\Gamma\Delta t}\boldsymbol p+\sqrt{1-e^{-2\Gamma\Delta t}}\sqrt{\frac{M}{\beta}}\boldsymbol{\xi}$

where $\boldsymbol{\xi}$ is sampled from standard normal distribution and $\Gamma$ is determined by the characteristic frequency.

PIMD.Normal — Type

Normal(quantum, n)

Creates an normal isomorphic PIMD system from a quantum system, with beads number n.

PIMD.OneDimensionalQuantumSystem — Type

OneDimensionalQuantumSystem(β, m, potential)

Creates an 1D quantum system with temperature β and Hamiltonian $H=p^2/2m+V(x)$.

PIMD.PotentialPropagator — Type

KineticPropagator(Δt)

Creates a propagator that updates momenta according to the canonical equation

$\boldsymbol p\leftarrow\boldsymbol p-\frac{\partial H}{\partial\boldsymbol q}\Delta t$

PIMD.Primitive — Type

Primitive(quantum, n)

Creates an primitive isomorphic PIMD system from a quantum system, with beads number n.

PIMD.Staging — Type

Staging(quantum, n)

Creates an staging isomorphic PIMD system from a quantum system, with beads number n.

PIMD.estimatorHessian — Function

estimatorHessian(phase, system)

Estimates the hessian for an isomorphic classical system under a phase.

$\hat H(\boldsymbol x)=\frac1n\sum_iV''(x_i)$

PIMD.estimatorKinetic — Function

estimatorKinetic(phase, system)

Estimates the kinetic energy (using the primitive estimator) for an isomorphic classical system under a phase.

$\hat K_{\rm p}(\boldsymbol x)=\frac n{2\beta}-\frac12\boldsymbol x^T\mathsf K\boldsymbol x$

PIMD.estimatorKineticVirial — Function

estimatorKinetic(phase, system)

Estimates the kinetic energy (using the virial estimator) for an isomorphic classical system under a phase.

$\hat K_{\rm v}(\boldsymbol x)=\frac 1{2\beta}+\frac1{2n}(\bar x-\boldsymbol x)\partial{\varphi(x)}{\boldsymbol x}$

PIMD.estimatorPotential — Function

estimatorPotential(phase, system)

Estimates the potential energy for an isomorphic classical system under a phase.

$\hat V(\boldsymbol x)=\varphi(\boldsymbol x)$

PIMD.md — Method

md(initPhase, system, integrator!, timetable, properties)

Launches an MD simulation of system with the initial conditions initPhase and collect properties. During the simulation, integrator! is used to integrate the equations of motion. With timetable = (equilibrationSteps, sampleSteps), the system is first equilibrated for equilibrationSteps and then sampled for sampleSteps.