API

Quantum System

PIMD.OneDimensionalQuantumSystem — Type

OneDimensionalQuantumSystem(β, m, potential)

Creates an 1D quantum system with temperature β and Hamiltonian $H=p^2/2m+V(x)$.

Classical System

PIMD.Primitive — Type

Primitive(quantum, n)

Creates an primitive isomorphic PIMD system from a quantum system, with beads number n.

PIMD.Staging — Type

Staging(quantum, n)

Creates an staging isomorphic PIMD system from a quantum system, with beads number n.

PIMD.Normal — Type

Normal(quantum, n)

Creates an normal isomorphic PIMD system from a quantum system, with beads number n.

Propagators

PIMD.KineticPropagator — Type

KineticPropagator(Δt)

Creates a propagator that updates positions according to the canonical equation

$\boldsymbol q\leftarrow\boldsymbol q+\frac{\partial H}{\partial\boldsymbol p}\Delta t$

PIMD.PotentialPropagator — Type

KineticPropagator(Δt)

Creates a propagator that updates momenta according to the canonical equation

$\boldsymbol p\leftarrow\boldsymbol p-\frac{\partial H}{\partial\boldsymbol q}\Delta t$

PIMD.ExternalPotentialPropagator — Type

ExternalKineticPropagator(Δt)

Creates a propagator that updates momenta according to the canonical equation, with PIMD external potential only

$\boldsymbol p\leftarrow\boldsymbol p-\frac{\partial V_{\rm ext}}{\partial\boldsymbol q}\Delta t$

PIMD.GeneralizedCayleyPropagator — Type

GeneralizedCayleyPropagator(system, generalizedCayleyFunction, Δt)

Creates a generalized Cayley propagator that PIMD internal potential using a generalized cayley function $\varphi$

$\boldsymbol z\leftarrow\varphi(\mathsf J\mathsf H)\boldsymbol z$

Examples of generalized Cayley functions are exp(), cayley() and sqrtcayley() (recommended).

PIMD.LangevinPropagator — Type

LangevinPropagator(system, Δt)

Creates a propagator that thermostats momenta using Langevin equation

$\boldsymbol p\leftarrow e^{-\Gamma\Delta t}\boldsymbol p+\sqrt{1-e^{-2\Gamma\Delta t}}\sqrt{\frac{M}{\beta}}\boldsymbol{\xi}$

where $\boldsymbol{\xi}$ is sampled from standard normal distribution and $\Gamma$ is determined by the characteristic frequency.