PIMD

PIMD.ExternalPotentialPropagator

PIMD.GeneralizedCayleyPropagator

PIMD.Integrator

PIMD.KineticPropagator

PIMD.LangevinPropagator

PIMD.Normal

PIMD.OneDimensionalQuantumSystem

PIMD.PotentialPropagator

PIMD.Primitive

PIMD.Staging

PIMD.estimatorHessian

PIMD.estimatorKinetic

PIMD.estimatorKineticVirial

PIMD.estimatorPotential

PIMD.md

PIMD.Integrator — Type

Integrator(propagators...)

Constructs an integrator from a series of propagator, so that within a time step these propagators act sequentially on the phase.

PIMD.estimatorHessian — Function

estimatorHessian(phase, system)

Estimates the hessian for an isomorphic classical system under a phase.

$\hat H(\boldsymbol x)=\frac1n\sum_iV''(x_i)$

PIMD.estimatorKinetic — Function

estimatorKinetic(phase, system)

Estimates the kinetic energy (using the primitive estimator) for an isomorphic classical system under a phase.

$\hat K_{\rm p}(\boldsymbol x)=\frac n{2\beta}-\frac12\boldsymbol x^T\mathsf K\boldsymbol x$

PIMD.estimatorKineticVirial — Function

estimatorKinetic(phase, system)

Estimates the kinetic energy (using the virial estimator) for an isomorphic classical system under a phase.

$\hat K_{\rm v}(\boldsymbol x)=\frac 1{2\beta}+\frac1{2n}(\bar x-\boldsymbol x)\partial{\varphi(x)}{\boldsymbol x}$

PIMD.estimatorPotential — Function

estimatorPotential(phase, system)

Estimates the potential energy for an isomorphic classical system under a phase.

$\hat V(\boldsymbol x)=\varphi(\boldsymbol x)$

PIMD.md — Method

md(initPhase, system, integrator!, timetable, properties)

Launches an MD simulation of system with the initial conditions initPhase and collect properties. During the simulation, integrator! is used to integrate the equations of motion. With timetable = (equilibrationSteps, sampleSteps), the system is first equilibrated for equilibrationSteps and then sampled for sampleSteps.